Removed code duplication
